:recycle: [Refactor] work, feedback 미들웨어 삭제
